backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
arq==0.26.3
python-multipart==0.0.6
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
pytest-asyncio==0.23.3
ruff>=0.14.0  # Linting and formatting
aiogoogle==5.11.0
youtube-transcript-api==0.6.3
tenacity==8.2.3
isodate==0.6.1
google-genai[aiohttp]==0.3.0
rapidfuzz==3.10.0  # High-performance Levenshtein distance
numpy==1.26.4  # For embedding operations
yt-dlp>=2024.07.01  # Video/audio downloading for enrichment (CVE fixes)
groq>=0.4.0  # Whisper transcription API
pydub>=0.25.0  # Audio processing for chunking
feedparser>=6.0.0  # Parse Atom/RSS feeds from PubSubHubbub
aiolimiter>=1.1.0
